:sparkles: chnange: move check bundle name to the Good Practices validator
